+++ include/smallship.h
@@ -1,6 +1,7 @@
 #ifndef DREV_SMALL_SHIP_H
 #define DREV_SMALL_SHIP_H
 
+#include "types.h"
 #include "vec.h"
 
 #include <SDL.h>
@@ -8,18 +9,18 @@
 typedef struct drev_small_ship SmallShip;
 
 struct drev_small_ship {
-    Vec2  position;
-    Vec2  target_position;
-    Vec2  velocity;
-    Vec2  size;
-    float rotation;
-    float fire_cooldown;
+    Vec2 position;
+    Vec2 target_position;
+    Vec2 velocity;
+    Vec2 size;
+    f32  rotation;
+    f32  fire_cooldown;
 };
 
-int       SmallShip_Init(SmallShip *ship);
-SmallShip SmallShip_Create(Vec2 position, Vec2 velocity, float rotation);
+i8        SmallShip_Init(SmallShip *ship);
+SmallShip SmallShip_Create(Vec2 position, Vec2 velocity, f32 rotation);
 void      SmallShip_MoveTo(SmallShip *ship, Vec2 new_position);
-void      SmallShip_Update(SmallShip *ship, float delta);
+void      SmallShip_Update(SmallShip *ship, f32 delta);
 SDL_FRect SmallShip_GetBounds(const SmallShip *ship);
 void      SmallShip_Draw(SmallShip *ship);
